Add date functions to a util
